+++ src/modules/outlierDetection/getIntersectionPoint.js
@@ -1,0 +1,3 @@
+import { intersect } from 'mathjs'
+
+export const getIntersectionPoint = (...points) => intersect(points)
