+++ src/modules/outlierDetection/getNormalEndPoint.js
@@ -1,0 +1,4 @@
+const normalInc = 1
+
+export const getNormalEndPoint = (dataPoint, normalGradient) =>
+[dataPoint[0] + normalInc, dataPoint[1] + normalInc * normalGradient]
