+++ src/modules/outlierDetection/getNormalGradient.js
@@ -1,0 +1,1 @@
+export const getNormalGradient = graphGradient => -1 / graphGradient
