+++ src/modules/outlierDetection/getNormalGraph.js
@@ -1,0 +1,4 @@
+import { getNormalEndPoint } from './getNormalEndPoint'
+
+export const getNormalGraph = (dataPoint, normalGradient) =>
+[dataPoint, getNormalEndPoint(dataPoint, normalGradient)]
