+++ src/modules/outlierDetection/getPointDistance.js
@@ -1,0 +1,3 @@
+import { distance } from 'mathjs'
+
+export const getPointDistance = (...points) => distance(...points)
