+++ src/modules/outlierDetection/getStdDev.js
@@ -1,0 +1,3 @@
+import { std } from 'mathjs'
+
+export const getStdDev = (...data) => std(data)
