+++ src/modules/outlierDetection/index.js
@@ -1,25 +1,28 @@
-import { getNormalGradient } from "./functions"
-import { getRegression, getStdDev, getNormalGraph, getPointDistance, getIntersectPoint } from "./utils"
+import { getStdDev } from './getStdDev'
+import { getRegression } from "./getRegression"
+import { getIntersectionPoint } from "./getIntersectionPoint"
+import { getPointDistance } from "./getPointDistance"
+import { getNormalGraph } from "./getNormalGraph"
+import { getNormalGradient } from './functions'
 
-export const detectOutliers = (data, stdDevThreshold = 1) => {
+export default (data, stdDevThreshold = 1) => {
     const stdDevValue = getStdDev(data) * stdDevThreshold
     const reg = getRegression(data)
     
     const regGradient = reg.equation[0]
-    const regIntersectGraph = [reg.points[0], reg.points[1]]
-
+    const regIntersectionGraph = [reg.points[0], reg.points[1]]
     const normalGradient = getNormalGradient(regGradient)
 
     const outlierPoints = []
     // const stdDevGraphs = []
 
-    let normalIntersectGraph
+    let normalIntersectionGraph
     let intersectPoint
     let pointDistance
 
     data.forEach(dataPoint => {
         normalIntersectGraph = getNormalGraph(dataPoint, normalGradient)
-        intersectPoint = getIntersectPoint([...normalIntersectGraph, ...regIntersectGraph])
+        intersectPoint = getIntersectionPoint([...normalIntersectionGraph, ...regIntersectionGraph])
         pointDistance = getPointDistance(dataPoint, intersectPoint)
 
         pointDistance > stdDevValue && (outlierPoints.push(dataPoint))
